Convert core.
