+++ packages/core/src/Markup.tsx
@@ -2,18 +2,8 @@
 
 import React from 'react';
 import Element from './Element';
-import Parser, { ParserProps } from './Parser';
-
-export interface MarkupProps extends ParserProps {
-  /** Content that may contain HTML to safely render. */
-  content?: string | null;
-  /** Content to render when the `content` prop is empty. */
-  emptyContent?: React.ReactNode;
-  /** @ignore Pre-parsed content to render. */
-  parsedContent?: React.ReactNode;
-  /** HTML element or React fragment to wrap content with. */
-  tagName?: 'fragment' | string;
-}
+import Parser from './Parser';
+import { MarkupProps } from './types';
 
 export default function Markup(props: MarkupProps) {
   const { content, emptyContent, parsedContent, tagName } = props;
