+++ packages/core/src/Parser.ts
@@ -2,10 +2,8 @@
 
 import React from 'react';
 import escapeHtml from 'escape-html';
-import Element, { ElementProps } from './Element';
-import { FilterInterface, ElementAttributes } from './Filter';
+import Element from './Element';
 import StyleFilter from './StyleFilter';
-import { MatcherInterface } from './Matcher';
 import {
   FILTER_DENY,
   FILTER_CAST_NUMBER,
@@ -21,9 +19,14 @@ import {
   Attributes,
   Node,
   NodeConfig,
-  TransformCallback,
   AttributeValue,
   ChildrenNode,
+  ParserProps,
+  MatcherElementsMap,
+  ElementProps,
+  FilterInterface,
+  ElementAttributes,
+  MatcherInterface,
 } from './types';
 
 const ELEMENT_NODE = 1;
@@ -31,35 +34,6 @@ const TEXT_NODE = 3;
 const INVALID_ROOTS = /^<(!doctype|(html|head|body)(\s|>))/i;
 const ALLOWED_ATTRS = /^(aria-|data-|\w+:)/iu;
 const OPEN_TOKEN = /{{{(\w+)\/?}}}/;
-
-interface MatcherElementsMap {
-  [key: string]: {
-    children: string;
-    matcher: MatcherInterface<{}>;
-    props: object;
-  };
-}
-
-export interface ParserProps {
-  /** Disable filtering and allow all non-banned HTML attributes. */
-  allowAttributes?: boolean;
-  /** Disable filtering and allow all non-banned/blocked HTML elements to be rendered. */
-  allowElements?: boolean;
-  /** List of HTML tag names to allow and render. Defaults to the `ALLOWED_TAG_LIST` constant. */
-  allowList?: string[];
-  /** List of HTML tag names to disallow and not render. Overrides allow list. */
-  blockList?: string[];
-  /** Disable the conversion of new lines to `<br />` elements. */
-  disableLineBreaks?: boolean;
-  /** Escape all HTML before parsing. */
-  escapeHtml?: boolean;
-  /** Strip all HTML while rendering. */
-  noHtml?: boolean;
-  /** Strip all HTML, except HTML generated by matchers, while rendering. */
-  noHtmlExceptMatchers?: boolean;
-  /** Transformer ran on each HTML element. Return a new element, null to remove current element, or undefined to do nothing. */
-  transform?: TransformCallback | null;
-}
 
 export default class Parser {
   allowed: Set<string>;
