+++ packages/core/src/index.ts
@@ -3,25 +3,14 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-import Interweave, { InterweaveProps } from './Interweave';
-import Markup, { MarkupProps } from './Markup';
-import Filter, { FilterInterface, ElementAttributes } from './Filter';
-import Matcher, { MatcherInterface } from './Matcher';
+import Interweave from './Interweave';
+import Markup from './Markup';
+import Filter from './Filter';
+import Matcher from './Matcher';
 import Element from './Element';
 import Parser from './Parser';
 
-export {
-  InterweaveProps,
-  Markup,
-  MarkupProps,
-  Filter,
-  FilterInterface,
-  ElementAttributes,
-  Matcher,
-  MatcherInterface,
-  Element,
-  Parser,
-};
+export { Markup, Filter, Matcher, Element, Parser };
 
 export * from './constants';
 export * from './types';
